optimized code remove unnecessary classes
