--- src/endToEndTest/java/org/radarcns/integration/commons/EndToEndUtility.java
@@ -1,53 +1,0 @@
-/*
- *  Copyright 2017 Kings College London and The Hyve
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.radarcns.integration.commons;
-
-import java.io.IOException;
-import java.util.concurrent.TimeUnit;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-/**
- * Created by nivethika on 23-3-17.
- */
-public class EndToEndUtility {
-    private static final Logger LOGGER = LoggerFactory.getLogger(EndToEndUtility.class);
-
-    /**
-     * Makes an HTTP request to given URL.
-     * @param url end-point
-     * @return HTTP Response
-     * @throws IOException
-     */
-    public static Response makeRequest(String url) throws IOException {
-        OkHttpClient client = new OkHttpClient.Builder()
-                .connectTimeout(30, TimeUnit.SECONDS)
-                .writeTimeout(30, TimeUnit.SECONDS)
-                .readTimeout(30, TimeUnit.SECONDS)
-                .build();
-
-        Request request = new Request.Builder()
-                .header("User-Agent", "Mozilla/5.0")
-                .url(url)
-                .build();
-
-        return client.newCall(request).execute();
-    }
-}
